% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{forecast}
\alias{forecast}
\title{Forecast from a DSGE Model}
\usage{
forecast(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A forecast object.
}
\description{
Generic function for forecasting from estimated DSGE models.
}
\seealso{
\code{\link[=forecast.dsge_fit]{forecast.dsge_fit()}}
}
