% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoother.R
\name{plot.dsge_smoothed}
\alias{plot.dsge_smoothed}
\title{Plot Smoothed States}
\usage{
\method{plot}{dsge_smoothed}(x, which = NULL, type = c("states", "fit"), ...)
}
\arguments{
\item{x}{A \code{dsge_smoothed} object.}

\item{which}{Which states to plot. Integer vector, character vector of
state names, or \code{NULL} (all states).}

\item{type}{Either \code{"states"} or \code{"fit"}. \code{"states"} plots the smoothed
state variables; \code{"fit"} plots the smoothed observables against data.}

\item{...}{Additional arguments passed to \code{\link[=plot]{plot()}}.}
}
\description{
Plot Smoothed States
}
